api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Cloud
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
firebase-admin==6.4.0

# Stripe
stripe==7.11.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
slowapi==0.1.9

# Image handling
Pillow==10.2.0
